core: enable eda-api svc nodePorts if no loadbalancer is used
Kubernetes implements services of LoadBalancer type by
requesting a service of NodePort type and then making changes
to it.

The current recommendation to deploy is metallb as a
LoadBalancer which does not require NodePorts on Services to
function. In security scans of a running system NodePorts are
reported as being open/vulnerable (without merit).

In 25.x it was decided that the default behaviour should be
to not create these NodePorts on the eda-api service with an
option to toggle if required.

Cloud Loadbalancers do require these NodePorts to route
traffic from their edge to the nodes where the scheduled pod
is running. Each cloud has its own variation on labels and
annotations to add if NodePorts are turned off on LB Services
for it to function.

In the Playground enable NodePort if NO_LB=1 is set to 1, ie
$USER is not using metallb and likely no LoadBalancer or their
cluster/cloud provided one.