git: scale down eda-git+replica deployments
When using a storage driver which specifies
nodeattachments for volumeclaims, if a deployment
has to be respun for that pod, the new pod will be
stuck in containerCreating state because it cannot
attach to the volume since it is already attached to
an existing pod, since the new pod doesn't start the
old pod is not scaled down.

This target scales down the pods in the eda-git and
eda-git-replica deployments which on installing a new
version of the package the new pod will spin up and
attach it self to the pv because no running pod is
already attached and we don't need to delete it.